Add initial point tests
