+++ geoarrow/shapely/geoarrow/shapely/extension_types.py
@@ -273,29 +273,6 @@ class MultiPolygonType(BaseGeometryType):
         super().__init__(storage_type, self.extension_name)
 
 
-def register_geometry_extension_types():
-    for geom_type_class in [
-        PointType,
-        LineStringType,
-        PolygonType,
-        MultiPointType,
-        MultiLineStringType,
-        MultiPolygonType,
-    ]:
-        # Provide a default to go into the registry, but at runtime, we can choose other
-        # type formulations
-        geom_type_instance = geom_type_class(
-            interleaved=True, dims=CoordinateDimension.XY
-        )
-        try:
-            pa.register_extension_type(geom_type_instance)
-
-        # If already registered with this id, unregister and re register
-        except pa.ArrowKeyError:
-            pa.unregister_extension_type(geom_type_instance.extension_name)
-            pa.register_extension_type(geom_type_instance)
-
-
 # register_geometry_extension_types()
 # shapely_arr = shapely.points([[1, 2], [3, 4], [5, 6], [5, 6]])
 # point_arr = construct_geometry_array(shapely_arr)
